+++ src/lib/st-search/st-search.component.spec.ts
@@ -8,7 +8,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0.
  */
-import { DebugElement, SimpleChanges, SimpleChange, ChangeDetectionStrategy } from '@angular/core';
+import { ChangeDetectionStrategy, DebugElement, SimpleChange, SimpleChanges } from '@angular/core';
 import { async, ComponentFixture, fakeAsync, TestBed, tick } from '@angular/core/testing';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { By } from '@angular/platform-browser';
@@ -17,6 +17,7 @@ import { StDropdownMenuModule } from '../st-dropdown-menu/st-dropdown-menu.modul
 import { StDropDownMenuItem } from '../st-dropdown-menu/st-dropdown-menu.interface';
 import { StSelectModule } from '../st-select/st-select.module';
 import { StInputModule } from '../st-input/st-input.module';
+import { StSearchEventOrigin } from './st-search.model';
 
 // Component
 
@@ -32,7 +33,7 @@ describe('StSearchComponent', () => {
 
    beforeEach(async(() => {
       TestBed.configureTestingModule({
-         imports: [FormsModule, ReactiveFormsModule, StDropdownMenuModule, StSelectModule, StInputModule],
+         imports: [FormsModule, ReactiveFormsModule, StDropdownMenuModule, StSelectModule, StInputModule, StDropdownMenuModule],
          declarations: [StSearchComponent]
       })
       // remove this block when the issue #12313 of Angular is fixed
@@ -82,7 +83,7 @@ describe('StSearchComponent', () => {
       fixture.detectChanges();
 
       tick(11);
-      expect(component.search.emit).toHaveBeenCalledWith({text: result});
+      expect(component.search.emit).toHaveBeenCalledWith({ text: result, origin: StSearchEventOrigin.INPUT });
    }));
 
    it('should be able to search with the delay introduced as input', fakeAsync(() => {
@@ -99,7 +100,7 @@ describe('StSearchComponent', () => {
       tick(500);
       expect(component.search.emit).not.toHaveBeenCalled();
       tick(501);
-      expect(component.search.emit).toHaveBeenCalledWith({text: result});
+      expect(component.search.emit).toHaveBeenCalledWith({ text: result, origin: StSearchEventOrigin.INPUT });
    }));
 
    it('should be able to search with an updated delay', fakeAsync(() => {
@@ -116,7 +117,7 @@ describe('StSearchComponent', () => {
       tick(500);
       expect(component.search.emit).not.toHaveBeenCalled();
       tick(501);
-      expect(component.search.emit).toHaveBeenCalledWith({text: result});
+      expect(component.search.emit).toHaveBeenCalledWith({ text: result, origin: StSearchEventOrigin.INPUT });
 
       let changes: SimpleChanges = { debounce: new SimpleChange(1000, 250, true) };
       component.debounce = 250;
@@ -132,7 +133,7 @@ describe('StSearchComponent', () => {
       tick(200);
       expect(component.search.emit).not.toHaveBeenCalled();
       tick(55);
-      expect(component.search.emit).toHaveBeenCalledWith({text: result});
+      expect(component.search.emit).toHaveBeenCalledWith({ text: result, origin: StSearchEventOrigin.INPUT });
    }));
 
    it('should search when user types a text with the min length introduced as input', fakeAsync(() => {
@@ -154,7 +155,7 @@ describe('StSearchComponent', () => {
 
       tick(1);
       expect(component.search.emit).toHaveBeenCalled();
-      expect(component.search.emit).toHaveBeenCalledWith({text: 'test'});
+      expect(component.search.emit).toHaveBeenCalledWith({ text: 'test', origin: StSearchEventOrigin.INPUT });
    }));
 
    it('should search when user presses the enter key', () => {
@@ -171,7 +172,7 @@ describe('StSearchComponent', () => {
 
       expect(component.search.emit).toHaveBeenCalled();
       expect(component.search.emit).toHaveBeenCalledTimes(1);
-      expect(component.search.emit).toHaveBeenCalledWith({text: 'te'});
+      expect(component.search.emit).toHaveBeenCalledWith({ text: 'te', origin: StSearchEventOrigin.ENTER });
 
       input.nativeElement.value = 'test';
       input.nativeElement.dispatchEvent(new Event('input'));
@@ -188,14 +189,14 @@ describe('StSearchComponent', () => {
 
       expect(component.search.emit).toHaveBeenCalled();
       expect(component.search.emit).toHaveBeenCalledTimes(2);
-      expect(component.search.emit).toHaveBeenCalledWith({text: 'test'});
+      expect(component.search.emit).toHaveBeenCalledWith({ text: 'test', origin: StSearchEventOrigin.ENTER });
 
       input.triggerEventHandler('keypress', { which: 13 });
       fixture.detectChanges();
 
       expect(component.search.emit).toHaveBeenCalled();
       expect(component.search.emit).toHaveBeenCalledTimes(3);
-      expect(component.search.emit).toHaveBeenCalledWith({text: 'test'});
+      expect(component.search.emit).toHaveBeenCalledWith({ text: 'test', origin: StSearchEventOrigin.ENTER });
    });
 
    it('should be able to search twice', fakeAsync(() => {
@@ -209,7 +210,7 @@ describe('StSearchComponent', () => {
       tick(1);
       expect(component.search.emit).toHaveBeenCalled();
       expect(component.search.emit).toHaveBeenCalledTimes(1);
-      expect(component.search.emit).toHaveBeenCalledWith({text: 'te'});
+      expect(component.search.emit).toHaveBeenCalledWith({ text: 'te', origin: StSearchEventOrigin.INPUT });
 
       input.nativeElement.value = 'te';
       input.nativeElement.dispatchEvent(new Event('input'));
@@ -225,7 +226,7 @@ describe('StSearchComponent', () => {
       tick(1);
       expect(component.search.emit).toHaveBeenCalled();
       expect(component.search.emit).toHaveBeenCalledTimes(2);
-      expect(component.search.emit).toHaveBeenCalledWith({text: 'test'});
+      expect(component.search.emit).toHaveBeenCalledWith({ text: 'test', origin: StSearchEventOrigin.INPUT });
    }));
 
    it('should search when user presses the enter key without liveSearch', () => {
@@ -246,7 +247,7 @@ describe('StSearchComponent', () => {
       fixture.detectChanges();
 
       expect(component.search.emit).toHaveBeenCalled();
-      expect(component.search.emit).toHaveBeenCalledWith({text: 'test'});
+      expect(component.search.emit).toHaveBeenCalledWith({ text: 'test', origin: StSearchEventOrigin.ENTER });
    });
 
    it('should be able to clean the search text when user clicks on the cross button and event is emitted with an empty search', () => {
@@ -268,7 +269,7 @@ describe('StSearchComponent', () => {
       clearButton.nativeElement.dispatchEvent(new Event('mousedown'));
 
       expect(component.searchBox.value).toEqual('');
-      expect(component.search.emit).toHaveBeenCalledWith({text: ''});
+      expect(component.search.emit).toHaveBeenCalledWith({ text: '', origin: StSearchEventOrigin.INPUT });
    });
 
    it('should be able to change initial value of search', () => {
@@ -281,7 +282,7 @@ describe('StSearchComponent', () => {
       expect((<HTMLInputElement>input.nativeElement).value).toEqual('Initial value');
 
       let changes: SimpleChanges = {
-         value: new SimpleChange('Initial value',  'new value', true),
+         value: new SimpleChange('Initial value', 'new value', true),
          liveSearch: new SimpleChange(true, false, true)
       };
 
@@ -332,7 +333,7 @@ describe('StSearchComponent', () => {
       expect(component.isActive).toBeFalsy();
    });
 
-   it('should be able to search with autocomplete', () => {
+   it('should be able to search with autocomplete clicking on an option', () => {
       component.autocompleteList = [{ value: '1', label: '1' }, { value: '2', label: '2' }];
       component.withAutocomplete = true;
 
@@ -345,7 +346,7 @@ describe('StSearchComponent', () => {
       component.changeOption(component.autocompleteList[0]);
       expect(component.search.emit).toHaveBeenCalled();
       expect(component.search.emit).toHaveBeenCalledTimes(1);
-      expect(component.search.emit).toHaveBeenCalledWith({text: component.autocompleteList[0].label});
+      expect(component.search.emit).toHaveBeenCalledWith({ text: component.autocompleteList[0].label, origin: StSearchEventOrigin.LIST });
    });
 
    it('should be able to search with autocomplete', fakeAsync(() => {
@@ -358,7 +359,6 @@ describe('StSearchComponent', () => {
 
       expect(input).toBeDefined();
       expect(dropdownItems.length).toEqual(0);
-
 
       input.value = 'te';
       input.dispatchEvent(new Event('input'));
@@ -453,7 +453,7 @@ describe('StSearchComponent', () => {
 
          fixture.nativeElement.querySelectorAll('.st-dropdown-menu-item')[1].click();
 
-         expect(component.search.emit).toHaveBeenCalledWith({text: '', filter: component.filterOptions[1].value});
+         expect(component.search.emit).toHaveBeenCalledWith({ text: '', filter: component.filterOptions[1].value, origin: StSearchEventOrigin.FILTER });
       });
    });
 });